Backend/requirements.txt
Django==4.2.23
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
python-dotenv==1.0.1
requests==2.32.3
PyJWT==2.8.0
pyotp==2.9.0